Refactor tests directory structure to mirror bumble package layout
Move profile, transport, and crypto unit tests into their corresponding
subdirectories under tests/ (tests/profiles, tests/transport,
tests/crypto, and tests/audio).

TAG=agy
CONV=1190a23d-d543-4238-82af-2fe005f957ad
