+++ tests/profiles/battery_service_test.py
