+++ tests/profiles/heart_rate_service_test.py
