+++ tests/profiles/mcp_test.py
